+++ Laborator6/src/Constants/constante.java
@@ -1,12 +1,12 @@
 package Constants;
 
-import SERVICE.Pisica;
-import model1.Animal;
+
+import Models.Animal;
 
 import java.util.List;
 import java.util.Objects;
 
-public class constante {
+public class Constante {
     public final String LISTA_ANIMALE_GOALA = "Niciun animal in lista";
 
     public static final String listavaccinurigoala = "Animalul nu are niciun vaccin";
