+++ Laborator6/src/Exceptions/Exceptie2.java
